requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
supabase-py==0.0.2
gunicorn==20.1.0
python-jose==3.3.0
pillow>=9.5.0
